+++ foo_spider_monkey_panel/panel_tooltip_param.h
@@ -12,7 +12,7 @@ struct PanelTooltipParam
     SIZE tooltipSize{};
 
     std::wstring fontName;
-    float fontSize{};
+    uint32_t fontSize{};
     uint32_t fontStyle{};
 };
